Trying to fix stack overflow
