+++ wright/src/grammar/parsers/expression.rs
@@ -6,9 +6,3 @@ pub mod underscore;
 
 /// Expression parser.
 pub mod expression;
-
-/// Binary expression parser.
-pub mod binary_expression;
-
-#[cfg(test)]
-mod binary_expression_tests;
